requirements.txt
pandas==1.2.3
jupyter==1.0.0
cx_Oracle==8.1.0
python-dotenv==0.17.0